+++ d2networking/d2netpacket/packet_add_player.go
@@ -12,17 +12,19 @@ type AddPlayerPacket struct {
 	Name      string                         `json:"name"`
 	X         int                            `json:"x"`
 	Y         int                            `json:"y"`
+	Act       int                            `json:"act"`
 	HeroType  d2enum.Hero                    `json:"hero"`
 	Equipment d2inventory.CharacterEquipment `json:"equipment"`
 	Stats     d2hero.HeroStatsState          `json:"heroStats"`
 }
 
-func CreateAddPlayerPacket(id, name string, x, y int, heroType d2enum.Hero, stats d2hero.HeroStatsState, equipment d2inventory.CharacterEquipment) NetPacket {
+func CreateAddPlayerPacket(id, name string, act, x, y int, heroType d2enum.Hero, stats d2hero.HeroStatsState, equipment d2inventory.CharacterEquipment) NetPacket {
 	return NetPacket{
 		PacketType: d2netpackettype.AddPlayer,
 		PacketData: AddPlayerPacket{
 			Id:        id,
 			Name:      name,
+			Act:       act,
 			X:         x,
 			Y:         y,
 			HeroType:  heroType,
