+++ d2networking/d2server/connection_manager.go
@@ -8,13 +8,16 @@ import (
 	"time"
 )
 
-// ConnectionManager is responsible for cleanup up connections accepted by the game server. As the server communicates over
-// UDP and is stateless we need to implement some loose state management via a ping/pong system. ConnectionManager also handles
+// ConnectionManager is responsible for cleanup up connections accepted by the
+// game server. As the server communicates over
+// UDP and is stateless we need to implement some loose state management via a
+// ping/pong system. ConnectionManager also handles
 // communication for graceful shutdowns.
 //
 // retries: # of attempts before the dropping the client
 // interval: How long to wait before each ping/pong test
-// gameServer: The *GameServer is argument provided for the connection manager to watch over
+// gameServer: The *GameServer is argument provided for the connection manager
+// to watch over
 // status: map of inflight ping/pong requests
 type ConnectionManager struct {
 	sync.RWMutex
@@ -50,7 +53,8 @@ func (c *ConnectionManager) Run() {
 func (c *ConnectionManager) checkPeers() {
 	for id, connection := range c.gameServer.clientConnections {
 		if connection.GetConnectionType() != d2clientconnectiontype.Local {
-			if err := connection.SendPacketToClient(d2netpacket.CreatePingPacket()); err != nil {
+			pingPacket := d2netpacket.CreatePingPacket()
+			if err := connection.SendPacketToClient(pingPacket); err != nil {
 				log.Printf("Cannot ping client id: %s", id)
 			}
 			c.RWMutex.Lock()
@@ -66,7 +70,8 @@ func (c *ConnectionManager) checkPeers() {
 	}
 }
 
-// Recv simply resets the counter, acknowledging we have received a pong from the client.
+// Recv simply resets the counter, acknowledging we have received a pong from
+// the client.
 func (c *ConnectionManager) Recv(id string) {
 	c.status[id] = 0
 }
@@ -81,8 +86,10 @@ func (c *ConnectionManager) Drop(id string) {
 
 // Shutdown will notify all of the clients that the server has been shutdown.
 func (c *ConnectionManager) Shutdown() {
-	// TODO: Currently this will never actually get called as the go routines are never signaled about the application termination.
-	// Things can be done more cleanly once we have graceful exits however we still need to account for other OS Signals
+	// TODO: Currently this will never actually get called as the go routines
+	// are never signaled about the application termination.
+	// Things can be done more cleanly once we have graceful exits however we
+	// still need to account for other OS Signals
 	log.Print("Notifying clients server is shutting down...")
 	for _, connection := range c.gameServer.clientConnections {
 		connection.SendPacketToClient(d2netpacket.CreateServerClosedPacket())
